+++ src/View/Template/ToolkitTemplateEngine.php
@@ -25,7 +25,7 @@ use Symfony\Component\Templating\TemplateReferenceInterface;
  *
  * @package Netzmacht\Contao\Toolkit\View\Template
  */
-class ToolkitTemplateEngine implements TemplateRenderer, EngineInterface
+class ToolkitTemplateEngine implements EngineInterface
 {
     /**
      * Template factory.